+++ index.ts
@@ -3,7 +3,7 @@ import { registerBgModule } from "./src/bg.js";
 import { JobManager } from "./src/manager.js";
 import { registerSubagentModule } from "./src/subagent.js";
 
-export default async function (pi: ExtensionAPI) {
+export default function (pi: ExtensionAPI) {
   const manager = new JobManager(pi);
   manager.init();